+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+# [1.163.0](https://github.com/betagouv/service-national-universel/compare/v1.162.0...v1.163.0) (2022-11-09)
+
+
+### Bug Fixes
+
+* **admin:** phone required for new ref structure ([#2013](https://github.com/betagouv/service-national-universel/issues/2013)) ([3606b22](https://github.com/betagouv/service-national-universel/commit/3606b2278174fd872271c56e3d62db057cacd782))
+* **admin:** young eligibility + situation ([bd980e8](https://github.com/betagouv/service-national-universel/commit/bd980e865c0dd216b6413d9158aa53221c7aa9fe))
+* **api:** allow null for fromPage in forms ([3ed5e7c](https://github.com/betagouv/service-national-universel/commit/3ed5e7c01259741ab7c0ca51326c0296db8b974f))
+* **api:** Force redirection to reinscription on set from admin ([c84058c](https://github.com/betagouv/service-national-universel/commit/c84058ce7af79bbdd535de99a9c5ae4efbd884d4))
+* **api:** Validate ID format Joi ([491c9e3](https://github.com/betagouv/service-national-universel/commit/491c9e30b6f3464da7363ee155fbfd7cc9eba845))
+* **app:** fix mobile upload v2 for new accounts ([6a935a8](https://github.com/betagouv/service-national-universel/commit/6a935a836227e1366ca8af6427e2b1df9764dc9b))
+* **app:** fix upload error handling v2 ([f990b87](https://github.com/betagouv/service-national-universel/commit/f990b878608dd8ec92dbbc9f04840076951e4798))
+* **app:** remove status change menus ([#2026](https://github.com/betagouv/service-national-universel/issues/2026)) ([2908d60](https://github.com/betagouv/service-national-universel/commit/2908d608a6b62a6d0216d40afa7aab5bb8ef55df))
+* **app:** upload error handling ([8b6dc98](https://github.com/betagouv/service-national-universel/commit/8b6dc989df8f6ca003e8509e279e429ba1aa72dd))
+* **lib:** Add terminale for grades ([e4164b0](https://github.com/betagouv/service-national-universel/commit/e4164b04cd570636903455eb3b63b4f1780cbddb))
+* **lib:** fix department mapping ([cf21ef9](https://github.com/betagouv/service-national-universel/commit/cf21ef913d380612232ed418e5021ed916077f0e))
+
+
+### Features
+
+* **admin:** Affiner les résultats des propositions de missions ([dc98904](https://github.com/betagouv/service-national-universel/commit/dc98904aaa0f7c44dce61face3e3d1853ae9e97f))
+* **app:** info message for file upload ([#2028](https://github.com/betagouv/service-national-universel/issues/2028)) ([79bda56](https://github.com/betagouv/service-national-universel/commit/79bda56587645096b257fcefb4ef68cae1cab580))
+
 # [1.162.0](https://github.com/betagouv/service-national-universel/compare/v1.161.0...v1.162.0) (2022-11-07)